Changing ROM initialization code to use $readmemh
This prevents compatibility problems with other FPGA software (such as Xilinx)

Replaced the ROM dump files from the original Parallax archives by ROM dumps that are compatible with $readmemh. These were provided by Ray Rodrick (Cluso99) in the old P8X32A_Emulation repo and were produced using a small Spin program which I also added. The program can be run on a regular Propeller and then on the P1V to verify that the hex dumps are identical.

Not all the ROM dump files are in use at this time, only the scrambled versions of the big blobs (rom_8000_bfff_font.hex and rom_c000_ffff_scrambled.hex). In a later update, I may reinstate the feature that enables and disables ROM and RAM in the various upper memory areas; for this I may need to other hex files which are really just partial dumps and contain the same information as the file names mentioned above.