Adding .QSF file for Parallax FPGA123-A9 board
This is the unmodified file from Chip Gracey's zip file, posted on March 14, 2016, about halfway down the page on http://forums.parallax.com/discussion/161545/propeller-1-2-3-fpga-board-60054-available-to-forum-members-375/p8
I only renamed the file to fit the existing naming scheme, I didn't add it to the Quartus project and I didn't try to compile it yet. Looks like it will be necessary to change the top level file, and either rename the pins to work together with the existing fpga123.v file, or create a new Verilog file that works with the pin names as specified in this QSF file.