Moving some more wires and regs around
This is needed in preparation for Xilinx support; the Xilinx tools don't like it when you use wires/regs before declaring them.

Also changed the scl register to be 20 bits instead of 32 to fix a warning about the unused 12 bits at the top

Also added the possibility to change the number of cogs to something less than 8.

Also added some statements to propagate the parameters between modules (particularly INVERT_COG_LEDS and the new NUMCOGS).