Initial support for Digilent Arty (Xilinx Artix-7 35T)
The Arty is a great FPGA board for makers but unfortunately it's only big enough to fit a Propeller with two cogs.

This was only tested to build successfully, but I haven't had a chance to test it on the hardware (that's why this is on a separate branch for now). Currently it's necessary to connect a Prop plug to JD with a couple of wires, but I think it's possible to modify the code a bit to use the on-board USB serial port as Prop Plug instead. Stay tuned!