Small modification to connect reset to CK_RST (only)
CK_RST (ChipKit Reset) is the reset line on the Arduino headers. There's also a pushbutton connected to it.
I connected an LED to the resn signal, confirming that the Propeller now resets when you push the button. Also apparently the first cog now starts running, but it appears that regardless of JP2, the DTR line doesn't influence the CK_RST line.