Re-implemented prop-plug emulation on Arty
This change makes it possible to download code to the P1V running on the Arty from the Propeller Tool via the on-board FTDI chip.

By default, the RXD/TXD lines of the chip are connected to P30/P31 on the Propeller, but they can be disconnected by setting SW0 to the ON position (=away from the edge of the PCB).

IMPORTANT: JP2 must be bridged to connect DTR on the FTDI chip to the reset pin of the P1V. The RESET button can also be used to reset the P1V.

I tried to create a simple circuit to make the RGB-LEDs not as eye-piercingly bright as they are, but when I did, it made the P1V stop working, so I commented it out.