Add constraints for PLL-derived clocks
A large amount of each cog is made up of registers that are
controlled/clocked by the per-cog PLL (VID, CNT registers, etc.), and
because these are logic-derived clocks in this design, these paths are
completely unconstrained, so tens of thousands of warnings show up in
the timing report.  Since these are supposed to run at anywhere up to
128MHz by spec, it makes sense to add constraints that declare this
clocks at the appropriate speeds, and also indicate that they are
asynchronous from each other as well as from the remainder of the cog
logic so that when used the paths will support the desired PLL speeds.

These changes to the constraints completely eliminate the tens of
thousands of warnings in the final timing report related to these
portions of the design.