Add asynchronous synchronizers for all Propeller input pins
All of the Propeller inputs are effectively asynchronous to the clocks driving the cogs.
In order to minimize the possibility of metastability, standard practice for asynchronous
data inputs to a flop is to pass them through a minimum of two, and preferably more, layers of
chained flops, clocked in the domain of the destination logic, at the expense of a few cycles of input latency.