Revise clocking strategy to make full use of Xilinx MMCM
Rather than use logic-based counters to generate divided clocks, do it with the Xilinx-specific blocks.
This aids in timing and ensures a strict phase relationship between all of the clocks.

Also modifies the video section to synthesize more cleanly by using the enable signal as a clock enable
on the vclk, rather than driving the clock net with a LUT that combines them with logic.