Initial version of top-level module for DE0-Nano for new architecture
The original targets provided by Parallax shared the same top-level file but I reorganized this a while ago so that each target has its own hardware-specific top file that refers to the new P1V.v file for everything that goes on in the Propeller.
This is the first attempt at porting the DE0-Nano target to this architecture.