Moved connection from cfg[7] to reset into dig.v where it belongs
The CLKSET assembly instruction has the ability to reset the Propeller by setting bit 7 to 1. In the original source files from Parallax, this was implemented in the top.v file which is somewhat of an odd architecture: yes, the cfg bits are generated as output from the dig (core) module and don't go up any higher so as a general rule it would make sense to handle cfg[7] in top.v (or p1v.v or whatever) but first of all, the output of the always@ goes straight back into the dig module, secondly it creates a requirement that all top modules implement this code, otherwise this feature wouldn't work.

So after this, the dig module contains the nres register (which was previously in all the "top" files), and the code to assign nres from cfg[7] combined with external resets is now contained in dig.v.

I purposely renamed the pin so that existing code doesn't compile without change. I also made the reset input active-high instead of active-low. I used the name "inp_res" to minimize the number of changes necessary in other modules, though I'm not 100% happy with having no reason for the "inp_" prefix which now looks like an inconsistent use of coding standards.

This commit also includes an updated Nexys4.xdc because it apparently has a reference to the reset logic, but I'm pretty sure the two changed lines that I fixed, were already broken before I renamed nres to inp_res, and I'm also pretty sure that there's a missing line under those lines that's supposed to deal with the fact that when running in RCSLOW mode, Vivado apparently can't close timing for the reset.