Modified DE0-Nano target to use the dig/core architecture
I created the P1V module once, because I mistakenly thought that there should be a hardware-independent module on top of everything in the old "top.v" module.

What I didn't realize is that the "dig.v" module can actually be regarded as the top module of the hardware-independent Propeller code, and the tim.v module replaces whatever is in the real Propeller to generate the cog and PLL clock from the crystal oscillator or the internal RCFAST and RCSLOW clock generators.

The architecture of the original source files was a little messy though, with an Altera-specific "altpll" as clock generator and the "tim" module to derive the cog clock and PLL clock. For the Xilinx targets we have a more elegant solution (based on code by Magnus Karlsson) where the clock generator and timing configuration module are combined in one module. This commit creates a similar Altera-specific timing module based on altpll and tim.v.

This commit implements the "dig" architecture on the DE0-Nano:
- Finished the altera_clock.sv module so it (sort of) matches the xilinx_clock.sv module
- Changed the DE0-Nano.sv file so it replaces the altpll clock and the tim module with the altera_clock module, and so it replaces the P1V module with the dig.v module.
- Changed the project file to add/remove the source files accordingly
- Modified the .sdc file to match the renamed pins, nets, clocks.