Style change: Use ~ for single-bit negation.
I found one instance where in a single line I used both ~ and !.  In Verilog, ! means boolean "false" rather than negation, which means for a multibit value the result produced by ~ versus ! is not necessarily the same.  Functionally this is no different in this case since both values are single bits, but there's no reason to be inconsistent.