Deleting features.v and the INVERT_COG_LEDS parameter
Inverting the cog LEDs should be (and is) done at the top level for all targets that need it (BeMicroCV and BeMicroCV-A9). The macro is from long ago when I was still a Verilog n00b :-)
The features.v include file was once intended to turn features on and off but this is also better done in the top level file (or creating a separate top level file altogether). And the only implemented feature was the inverting of the cog LEDs. So it would be empty now anyway.